requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
openai
pypdf2
python-dotenv
pydantic
pydantic-settings
numpy
tiktoken